+++ .vscode/settings.json
@@ -56,5 +56,5 @@
         "temp/": true,
         "Temp/": true
     },
-    "dotnet.defaultSolution": "UnderControl.sln"
+    "dotnet.defaultSolution": "Under-Control.sln"
 }
