+++ Assets/!.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b0450381bb50784baa03862d828b118
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
