+++ Assets/!/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06c249305785b61439bf6800423ff6bd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
