+++ Assets/!/Animation/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ed3b4497e51de4a44862eb876243de1e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
