+++ Assets/!/Materials/playerPlaceholder.mat.meta
