+++ Assets/!/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6457c67f94f1d5840a79bab14f8631e8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
