+++ Assets/!/Settings/InputSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0c41667727793c944911049e03dfd300
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
