+++ Assets/!/Shaders.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d38d9477ea7ab1409e24117e8ae85e0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
