+++ Assets/!/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 68ec810ef3c41184486907fe85c90fab
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
