+++ Assets/!/Textures/playerPlaceholder.png.meta
